requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
Pillow
seed
coverage
django-cors-headers
psycopg2-binary
python-dotenv 
psycopg2
djangorestframework 
drf_spectacular
pillow 
djangorestframework-simplejwt
httpx
google-generativeai
